feat: add useCallback and remove unneccessary comments
