+++ frontend/package.json
@@ -9,6 +9,7 @@
     "@emotion/styled": "^11.0.0",
     "@opengovsg/formsg-sdk": "^0.9.0",
     "axios": "^0.21.4",
+    "comlink": "^4.3.1",
     "country-flag-icons": "^1.2.10",
     "date-fns": "^2.23.0",
     "focus-visible": "^5.2.0",
