+++ frontend/src/typings/worker-loader.d.ts
@@ -4,7 +4,5 @@ declare module 'worker-loader!*' {
     constructor()
   }
 
-  // Uncomment this if you set the `esModule` option to `false`
-  // export = WebpackWorker;
   export default WebpackWorker
 }
